弾幕シューティングのソースコードを更新

+++ Chapter10/2/DrawUtil.cpp
@@ -18,26 +18,26 @@ void DrawUtil::DrawGraph(int x, int y, int psH, int tex, bool alpha, int secondT
 		v.pos.z = 0.1f;
 	}
 	//ç∂è„
-	vertices[0].pos.x = x;
-	vertices[0].pos.y = y;
+	vertices[0].pos.x = static_cast<float>(x);
+	vertices[0].pos.y = static_cast<float>(y);
 	vertices[0].u = 0.0f;
 	vertices[0].v = 0.0f;
 
 	//âEè„
-	vertices[1].pos.x = x+gw;
-	vertices[1].pos.y = y;
+	vertices[1].pos.x = static_cast<float>(x+gw);
+	vertices[1].pos.y = static_cast<float>(y);
 	vertices[1].u = 1.0f;
 	vertices[1].v = 0.0f;
 
 	//ç∂â∫
-	vertices[2].pos.x = x;
-	vertices[2].pos.y = y + gh;
+	vertices[2].pos.x = static_cast<float>(x);
+	vertices[2].pos.y = static_cast<float>(y + gh);
 	vertices[2].u = 0.0f;
 	vertices[2].v = 1.0f;
 
 	//âEâ∫
-	vertices[3].pos.x = x + gw;
-	vertices[3].pos.y = y + gh;
+	vertices[3].pos.x = static_cast<float>(x + gw);
+	vertices[3].pos.y = static_cast<float>(y + gh);
 	vertices[3].u = 1.0f;
 	vertices[3].v = 1.0f;
 
@@ -46,6 +46,6 @@ void DrawUtil::DrawGraph(int x, int y, int psH, int tex, bool alpha, int secondT
 		SetUseTextureToShader(1, tex);
 	}
 	SetUsePixelShader(psH);
-	DrawPrimitive2DToShader(vertices.data(), vertices.size(),
+	DrawPrimitive2DToShader(vertices.data(), static_cast<int>(vertices.size()),
 							DX_PRIMTYPE_TRIANGLESTRIP);
 }
