+++ Chapter10/2/Scene/CommandListScene.cpp
@@ -30,8 +30,9 @@ void CommandListScene::Draw()
 		0xffaaff, true);
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 	//紫
-	DrawBoxAA(margin_size, margin_size,
-		wsize.width - margin_size, wsize.height - margin_size,
+	DrawBoxAA(
+		static_cast<float>(margin_size), static_cast<float>(margin_size),
+		static_cast<float>(wsize.width - margin_size), static_cast<float>(wsize.height - margin_size),
 		0xff00ff, false, 3.0f);
 	DrawString(margin_size + 10, margin_size + 10, L"コマンド表", 0x008800);
 }
