+++ Chapter10/2/Scene/OverScene.cpp
@@ -46,7 +46,7 @@ void OverScene::FadeDraw()
 	float rate = static_cast<float>(frame_) /
 					static_cast<float>(fade_interval);	
 	DrawRotaGraph(320, 240, 1.0f, 0.0f, imgH_, true);
-	SetDrawBlendMode(DX_BLENDMODE_ALPHA, rate * 255);
+	SetDrawBlendMode(DX_BLENDMODE_ALPHA, static_cast<int>(rate * 255.0f));
 	DrawBox(0, 0, 640, 480, 0x000000, true);
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
